Refactor language code interfaces and entities.
